3d window with basic movement and teleport support

+++ code/server/src/main/java/com/hamming/halbo/model/UserLocation.java
@@ -7,9 +7,9 @@ public class UserLocation extends BasicObject{
     private Continent continent;
     private City city;
     private Baseplate baseplate;
-    private int x;
-    private int y;
-    private int z;
+    private double x;
+    private double y;
+    private double z;
 
     public UserLocation(HalboID id) {
         super(id);
@@ -55,27 +55,27 @@ public class UserLocation extends BasicObject{
         this.baseplate = baseplate;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
 
-    public int getZ() {
+    public double getZ() {
         return z;
     }
 
-    public void setZ(int z) {
+    public void setZ(double z) {
         this.z = z;
     }
 }
